requirements.txt
aiohttp==3.11.11
cdsapi==0.7.7
earthengine-api==1.7.13
folium==0.20.0
fsspec==2026.2.0
geopandas==1.1.2
loguru==0.7.3
oci==2.139.0
pyproj==3.7.2
PyYAML==6.0.3
rasterio==1.4.4
Requests==2.32.5
Shapely==2.1.2
streamlit==1.54.0
streamlit_file_browser==3.2.22
streamlit_folium==0.26.1
tqdm==4.66.5


